10 problem has been fixed

+++ 10/main.cpp
@@ -1,21 +1,19 @@
+#include <cstdint>
 #include <iostream>
 #include <map>
 #include <string>
-
-// решение на C++ по каким-то причинам не принимается, но
-// идентичное на Go проходит без проблем.
 
 int main()
 {
 	std::string s;
 	std::cin >> s;
-	std::map<char, int> counts;
-	int n = s.length();
-	int left = 0;
-	int right = n - 1;
+	std::map<char, uint64_t> counts;
+	uint64_t n = s.length();
+	uint64_t left = 0;
+	uint64_t right = n - 1;
 	while (left <= right)
 	{
-		int count = (n - left) * (left + 1);
+		uint64_t count = (n - left) * (left + 1);
 		counts[s[left]] += count;
 		if (left != right)
 			counts[s[right]] += count;
